+++ scripts/build.js
@@ -1,8 +1,47 @@
+/* eslint-disable @typescript-eslint/no-var-requires */
 const fs = require("fs");
 const path = require("path");
+const { execSync } = require("child_process");
 const { build } = require("esbuild");
 const prettier = require("prettier");
 const { dependencies } = require("../package.json");
+
+const externalPlugin = {
+	name: "external-plugin",
+	setup(build) {
+		let filter = /^[^./]|^\.[^./]|^\.\.[^/]/;
+		build.onResolve({ filter }, (args) => ({ path: args.path, external: true }));
+	}
+};
+
+const noCommentsPlugin = {
+	name: "no-comments-plugin",
+	setup(build) {
+		build.onEnd((result) => {
+			if (result.errors.length > 0) return;
+
+			const content = fs.readFileSync("index.js", "utf8");
+			fs.writeFileSync("index.js", content.replace(/\n?\/\/.*\n/g, ""));
+		});
+	}
+};
+
+const declarationPlugin = {
+	name: "declaration-plugin",
+	setup(build) {
+		build.onEnd((result) => {
+			if (result.errors.length > 0) return;
+			execSync("tsc");
+
+			if (!fs.existsSync("types/global")) fs.mkdirSync("types/global");
+			fs.readdirSync("src/lib/global").forEach((file) => {
+				fs.copyFileSync(`src/lib/global/${file}`, `types/global/${file}`);
+			});
+			const content = fs.readFileSync("types/index.d.ts", "utf8");
+			fs.writeFileSync("types/index.d.ts", content.replace("../src/lib/global/", "global/"));
+		});
+	}
+};
 
 const formatOptions = {
 	printWidth: 200,
@@ -15,73 +54,53 @@ const formatOptions = {
 
 const builds = [
 	{
-		name: "main",
+		name: "bin",
 		options: {
-			entryPoints: ["src/index.ts"],
-			outfile: "bin/index.js",
+			entryPoints: ["src/bin/index.ts"],
+			outfile: "bin.js",
 			bundle: true,
-			external: [...Object.keys(dependencies), "terser-webpack-plugin"],
+			external: [...Object.keys(dependencies)],
 			platform: "node",
 			banner: { js: "#!/usr/bin/env node\n" }
-		},
-		format: () => {
-			const filePath = path.resolve(process.cwd(), "bin/index.js");
-			const content = fs.readFileSync(filePath, "utf8");
-			const formatted = prettier.format(content, formatOptions);
-			fs.writeFileSync(filePath, formatted);
 		}
 	},
 	{
-		name: "loaders",
+		name: "lib",
 		options: {
-			entryPoints: fs.readdirSync("./src/loaders").map((file) => `./src/loaders/${file}`),
-			outdir: "bin/loaders"
-		},
-		format: () => {
-			const files = fs.readdirSync("./bin/loaders");
-			for (let i = 0; i < files.length; i++) {
-				const filePath = path.resolve(process.cwd(), `./bin/loaders/${files[i]}`);
-				const content = fs.readFileSync(filePath, "utf8");
-				const formatted = prettier.format(content, formatOptions);
-				fs.writeFileSync(filePath, formatted);
-			}
-		}
-	},
-	{
-		name: "modules",
-		options: {
-			entryPoints: fs.readdirSync("./src/modules").map((file) => `./src/modules/${file}`),
-			outdir: "bin/modules"
-		},
-		format: () => {
-			const files = fs.readdirSync("./bin/modules");
-			for (let i = 0; i < files.length; i++) {
-				const filePath = path.resolve(process.cwd(), `./bin/modules/${files[i]}`);
-				const content = fs.readFileSync(filePath, "utf8");
-				const formatted = prettier.format(content, formatOptions);
-				fs.writeFileSync(filePath, formatted);
-			}
+			entryPoints: ["src/lib/index.ts"],
+			outfile: "index.js",
+			bundle: true,
+			platform: "node",
+			format: "esm",
+			plugins: [externalPlugin, declarationPlugin, noCommentsPlugin]
 		}
 	}
 ];
 
+const format = (file) => {
+	const filePath = path.resolve(process.cwd(), file);
+	const content = fs.readFileSync(filePath, "utf8");
+	const formatted = prettier.format(content, formatOptions);
+	fs.writeFileSync(filePath, formatted);
+};
+
 if (process.argv.slice(2)[0] === "watch") {
-	for (let i = 0; i < builds.length; i++) {
-		const { name, options, format } = builds[i];
+	for (const currBuild of builds) {
+		const { name, options } = currBuild;
 		options.watch = {
 			onRebuild(error) {
 				if (error) return;
-				format();
+				format(options.outfile);
 				console.log(`Built ${name}`);
 			}
 		};
 	}
 }
 
-for (let i = 0; i < builds.length; i++) {
-	const { name, options, format } = builds[i];
+for (const currBuild of builds) {
+	const { name, options } = currBuild;
 	build(options).then(() => {
-		format();
+		format(options.outfile);
 		console.log(`Built ${name}`);
 	});
 }