+++ src/bin/plugins/text.ts
@@ -1,0 +1,12 @@
+interface TextOptions {
+	regex: RegExp;
+}
+
+export default function text(options: TextOptions = { regex: /\.txt$/ }) {
+	return {
+		name: "text",
+		transform(code: string, id: string) {
+			if (options.regex.test(id)) return `export default ${JSON.stringify(code)};`;
+		}
+	};
+}