+++ src/lib/index.ts
@@ -1,0 +1,6 @@
+// eslint-disable-next-line @typescript-eslint/triple-slash-reference
+/// <reference path="./global/index.d.ts" />
+
+export * from "./settings";
+export * from "./strings";
+export * from "./webpackutils";