+++ types/index.d.ts
@@ -1,4 +1,0 @@
-/// <reference path="./global.d.ts" />
-
-export { default as SettingsManager } from "../lib/settings";
-export { default as StringsManager } from "../lib/strings";