+++ CHANGELOG.md
@@ -1,12 +1,43 @@
-# v2.1.4
+# v3.0.0
+
+**BundleBD v3.0.0 has been basically rewritten, so many things will have changed that may not be included in the changelog. Check the [Wiki](https://github.com/Neodymium7/BundleBD/wiki) for more detailed, up to date documentation.**
 
 ## Bundler
 
--   [Added better support for Zlibrary config options](https://github.com/Neodymium7/BundleBD/pull/1)
+### Command and Configuration
 
-## General
+-   Options can now be supplied in the command arguments in a typical `--option value` format.
+-   The Bundler Configuration file is now a `bundlebd.config.js` file in the root of the project that exports an object or function with additional options.
+-   The `devOutput` and `readmeOutput` options have been removed, and the `entry` option has been renamed to `input`.
 
--   Fixed some more minor issues with typings
+### Plugin Configuration
+
+-   The Plugin Configuration file is now a `plugin.json` file in the plugin's root directory.
+-   The `meta` option has been removed in favor of placing meta options in the main JSON object.
+-   The `changelog` and `readme` options have been removed.
+-   The `zlibrary` option can now take a ZeresPluginLibrary config object with changelog, default settings, etc.
+-   Added `installScript` option to add a script in the bundled plugin that will install it for the user if run directly.
+
+### General
+
+-   Switched to Rollup instead of Webpack for cleaner output.
+-   As a result, more CSS Preprocessors are supported, and some more new options have been added. Some things may behave slightly differently.
+-   The bundler will now automatically create a bound instance of BdApi.
+-   Now uses [Zerthox's BdApi types](https://github.com/Zerthox/betterdiscord-types) instead of the old (mediocre) ones.
+
+## Library
+
+### Updated
+
+-   Removed `SettingsManager` and `StringsManager` in favor of the new `createSettings` and `createStrings` functions.
+
+-   New `createSettings` and `createStrings` functions behave similarly to previous iterations, but now have improved typings, and place the settings/strings properties directly on the returned object instead of through `get` and `set` methods.
+
+-   Removed previously deprecated `useSettingState` method from Settings object.
+
+### Added
+
+-   New `WebpackUtils` object with some helpful filters and utilities for finding Discord's internal modules. See the documentation for more information on each specific item.
 
 # v2.1.3